+++ amorphie.contract.core/Entity/Contract/ContractDefinitionLanguageDetail.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using amorphie.contract.core.Entity.Base;
 using amorphie.contract.core.Entity.Contract;
 
-namespace amorphie.contract.core;
+namespace amorphie.contract.core.Entity.Contract;
 
 [Table("ContractDefinitionLanguageDetail", Schema = "Cont")]
     public class ContractDefinitionLanguageDetail : AudiEntity
