+++ amorphie.contract.zeebe/Services/ExtensionService.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using amorphie.contract.zeebe.Services.Interfaces;
 
 namespace amorphie.contract.zeebe.Services
 {
