--- amorphie.contract.zeebe/Services/Interfaces/IDocumentDefinitionService.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using amorphie.contract.core.Entity.Common;
-using amorphie.contract.core.Entity.Document;
-using amorphie.contract.zeebe.Model.DocumentDefinitionDataModel;
-
-namespace amorphie.contract.zeebe.Services.Interfaces
-{
-    public interface IDocumentDefinitionService
-    {
-        Task<DocumentDefinition> DataModelToDocumentDefinition(dynamic documentDefinitionDataDynamic, Guid id);
-    }
-}
