+++ amorphie.contract/Validator/Contract/ContractDefinitionLanguageDetailValidator.cs
@@ -1,0 +1,12 @@
+﻿
+using amorphie.contract.core.Entity.Contract;
+using FluentValidation;
+
+public sealed class ContractDefinitionLanguageDetailValidator : AbstractValidator<ContractDefinitionLanguageDetail>
+    {
+        public ContractDefinitionLanguageDetailValidator()
+        {
+            // RuleFor(x => x.Name).NotNull();
+            // RuleFor(x => x.Name).MinimumLength(10);
+        }
+    }
